# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Kamu Web Audit API
  version: 1.0.0
  description: |
    Public API for the Kamu Web Audit platform (Nessus-like web application security scanner).
    This specification focuses on Targets, Scans, Findings, Rules Catalog and Auth Profiles.
servers:
  - url: https://api.kamu-web-audit.example.com
    description: Production
  - url: https://staging-api.kamu-web-audit.example.com
    description: Staging
...
paths:
...
  /api/audits:
    post:
      tags: [Scans]
      summary: Create a new audit run
      description: |
        Creates a new audit run for a given target URL.
        This operation is idempotent when an Idempotency-Key header is provided.
      operationId: createAudit
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuditRunRequest'
      responses:
        '201':
          description: Audit run created
          headers:
            Location:
              description: URL of the created audit run
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRunDetail'
        '200':
          description: Existing audit run returned for the same Idempotency-Key and request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRunDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional idempotency key for safely retrying POST /api/audits requests.
        When provided, the server will return the same audit run for duplicate
        requests with the same key, user and request payload. If the same key is
        reused with a different payload, the server will respond with 409 Conflict.
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    Conflict:
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds to wait before retrying the request
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    InternalError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      description: RFC 7807 problem+json response
      properties:
        type:
          type: string
          format: uri
          example: "https://api.kamu-web-audit.example.com/problems/validation-error"
        title:
          type: string
          example: "One or more validation errors occurred."
        status:
          type: integer
          format: int32
          example: 400
        detail:
          type: string
        instance:
          type: string
          example: "/scans/123"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      required:
        - title
        - status
...
security:
  - bearerAuth: []